refactor: move plugin code into plugin.py
